pyproject.toml
[tool.poetry]
name = "swole"
version = "0.1.0"
description = "Get your swole on!"
authors = ["te25son <te25son@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
psycopg2-binary = "^2.9.3"
returns = "^0.19.0"
mypy = "^0.960"
django-stubs = "<1.11.0"
django-widget-tweaks = "^1.4.12"
pytest-django = "^4.5.2"
Faker = "^13.12.0"
pytest-factoryboy = "^2.3.1"
pytest-xdist = "^2.5.0"
django-crispy-forms = "^1.14.0"
django-allauth = "^0.50.0"
debugpy = "^1.6.0"
black = "^22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"
environs = {extras = ["django"], version = "^9.5.0"}

[tool.poetry.dev-dependencies]
pylint = "^2.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
